requirement.txt
bcrypt==3.1.6
blinker==1.4
bokeh==0.12.5
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
numpy==1.16.3
opencv-python==4.1.0.25
packaging==19.0
pandas==0.24.2
Pillow==6.0.0
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
tornado==6.0.2
urllib3==1.25.2
Werkzeug==0.15.4
WTForms==2.2.1
